+++ src/main/java/blackjack/view/OutputView.java
@@ -6,7 +6,7 @@ import blackjack.domain.gamer.Name;
 import blackjack.dto.DealerDto;
 import blackjack.dto.PlayerDto;
 import blackjack.dto.PlayersDto;
-import blackjack.view.mapper.SuitMapper;
+
 import java.util.List;
 import java.util.Map;
 import java.util.StringJoiner;
